requirements.txt
azure-ai-projects>=1.0.0b7
azure-ai-agents>=1.0.0b2
azure-identity>=1.17.0
python-dotenv>=1.0.1
mcp>=1.12.0
# Optional async HTTP client (Agents async)
aiohttp>=3.10.0
httpx>=0.27.0
# A2A protocol SDK for AgentCard/AgentSkill and A2A client
a2a-sdk>=0.1.5
# Web framework for A2A servers
starlette>=0.37.2
uvicorn>=0.30.3
